new: - package domain - exception handler refac: - move classe to domain
