+++ src/main/java/med/voll/api/controller/MedicController.java
@@ -1,7 +1,10 @@
 package med.voll.api.controller;
 
 import jakarta.validation.Valid;
-import med.voll.api.medic.*;
+import med.voll.api.domain.medic.DataListMedic;
+import med.voll.api.domain.medic.Medic;
+import med.voll.api.domain.medic.MedicRepository;
+import med.voll.api.domain.medic.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -10,8 +13,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("medic")
