+++ src/main/java/med/voll/api/domain/adress/Adress.java
@@ -1,4 +1,4 @@
-package med.voll.api.adress;
+package med.voll.api.domain.adress;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
