+++ src/main/java/med/voll/api/domain/adress/AdressData.java
@@ -1,4 +1,4 @@
-package med.voll.api.adress;
+package med.voll.api.domain.adress;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
