+++ src/main/java/med/voll/api/domain/medic/DataDetailMedic.java
@@ -1,6 +1,6 @@
-package med.voll.api.medic;
+package med.voll.api.domain.medic;
 
-import med.voll.api.adress.Adress;
+import med.voll.api.domain.adress.Adress;
 
 public record DataDetailMedic(Long id, String name, String email, Speciality speciality, String crm, Adress adress) {
