+++ src/main/java/med/voll/api/domain/medic/MedicRepository.java
@@ -1,8 +1,7 @@
-package med.voll.api.medic;
+package med.voll.api.domain.medic;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Range;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MedicRepository extends JpaRepository<Medic, Long> {
