+++ src/main/java/med/voll/api/domain/patient/DataListPatient.java
@@ -1,4 +1,4 @@
-package med.voll.api.patient;
+package med.voll.api.domain.patient;
 
 public record DataListPatient(String name, String phoneNumber) {
