+++ src/main/java/med/voll/api/domain/patient/DataUpdatePatient.java
@@ -1,4 +1,4 @@
-package med.voll.api.patient;
+package med.voll.api.domain.patient;
 
 public record DataUpdatePatient(Long id,
