+++ src/main/java/med/voll/api/domain/patient/PatientRepository.java
@@ -1,4 +1,4 @@
-package med.voll.api.patient;
+package med.voll.api.domain.patient;
 
 import org.springframework.data.jpa.repository.JpaRepository;
