+++ src/main/java/med/voll/api/infra/ExceptionDTO.java
@@ -1,0 +1,4 @@
+package med.voll.api.infra;
+
+public record ExceptionDTO(String message, String statusCode) {
+}
