+++ src/main/resources/application.properties
@@ -13,3 +13,5 @@ server.port=8080
 # Flyway settings
 spring.flyway.baseline-on-migrate=true
 
+server.error.include-stacktrace=never
+
